reorganized files again, generic highlight loader
