+++ src/app/content/highlights/hooks/initializeMyHighlightsSummary.ts
@@ -4,8 +4,8 @@ import { actionHook, assertDefined } from '../../../utils';
 import { summaryPageSize } from '../../constants';
 import * as selectContent from '../../selectors';
 import { highlightLocationFilters } from '../../selectors';
-import { extractTotalCounts } from '../../utils/sharedHighlightsUtils';
 import { initializeMyHighlightsSummary, receiveHighlightsTotalCounts, receiveSummaryHighlights } from '../actions';
+import { extractTotalCounts } from '../utils/summaryHighlightsUtils';
 import { loadMoreMyHighlights } from './loadMore';
 
 export const hookBody: ActionHookBody<typeof initializeMyHighlightsSummary> = (services) => async() => {
