+++ src/app/content/highlights/utils/getHighlightLocationFilterForPage.ts
