+++ src/app/content/highlights/utils/index.ts
@@ -1,6 +1,7 @@
 export { default as getHighlightLocationFilters } from './getHighlightLocationFilters';
 export { default as getHighlightLocationFiltersWithContent } from './getHighlightLocationFiltersWithContent';
 export { default as getHighlightColorFiltersWithContent } from './getHighlightColorFiltersWithContent';
+export { default as getHighlightLocationFilterForPage } from './getHighlightLocationFilterForPage';
 
 export {
   addToTotalCounts,
